Fix key-frame only mode after determining frame count
The iterator returned by packets() actually advances a cursor in the
input video file. After the frames have been counted the cursor is
at the end of the file causing any further calls to packets() to
return an iterator that immediately yeilds None.

The fix is to seek back to the start of the input file before creating
the second iterator.